pyproject.toml
[project]
name = "staff-echo"
version = "0.1.0"
description = "Staff-Echo AI Chatbot — customer-facing chatbot trained on staff voice data"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "websockets>=11.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
google = [
    "google-cloud-bigquery>=3.10.0",
    "google-cloud-speech>=2.20.0",
    "google-generativeai>=0.3.0",
]
mcp = [
    "mcp>=1.0.0",
]
redis = [
    "redis>=5.0.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["staff_echo/tests"]

[tool.coverage.run]
source = ["staff_echo"]
omit = ["tests/*"]
